Rename account to custom_account
