+++ revm/custom_account/management/commands/seed_superuser.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
-from account.models import CustomUser
+from custom_account.models import CustomUser
 
 
 class Command(BaseCommand):
